feat: Implement removing build directory

+++ util-iso.sh
@@ -212,9 +212,14 @@ run_build() {
         fi
     done
     show_elapsed_time "${FUNCNAME}" "${timer_start}"
-    if [ "$work_dir"="/tmp/cachyos-live-iso-tmp" ]; then
+    if [[ "$work_dir"="/tmp/cachyos-live-iso-tmp" ]] && [[ "$remove_build_dir" == "false" ]]; then
         msg "!!! Remember to remove $work_dir !!!"
         msg2 "sudo rm -rf $work_dir"
+    fi
+    if [[ "$remove_build_dir" == "true" ]]; then
+        msg "Automatically removing build directory ($work_dir)..."
+        rm -rf $work_dir
+        msg2 "Removed"
     fi
 }
